some changes on hannah file uplaod
