--- src/pages/project/index.vue
@@ -1,9 +1,0 @@
-<template>
-  <ProjectTable />
-  <Breadcrumbs />
-</template>
-
-<script setup>
-import ProjectTable from '@/components/ProjectTable.vue'
-import Breadcrumbs from '@/components/wazeef/Breadcrumbs.vue'
-</script>
