+++ src/pages/upload-cv/employee-info/index.vue
@@ -1,0 +1,7 @@
+<template>
+  <EmployeeInfoForm />
+</template>
+
+<script setup>
+import EmployeeInfoForm from '@/components/EmployeeInfoForm.vue'
+</script>
